+++ fe/fe-core/pom.xml
@@ -793,6 +793,15 @@ under the License.
             <artifactId>threeten-extra</artifactId>
             <version>1.7.2</version>
         </dependency>
+
+        <!-- https://mvnrepository.com/artifact/org.mockito/mockito-core -->
+        <dependency>
+            <groupId>org.mockito</groupId>
+            <artifactId>mockito-inline</artifactId>
+            <version>4.11.0</version>
+            <scope>test</scope>
+        </dependency>
+
     </dependencies>
 
     <build>
