+++ fe/fe-core/src/main/java/com/starrocks/load/pipe/filelist/FileListRepo.java
@@ -12,7 +12,10 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-package com.starrocks.load.pipe;
+package com.starrocks.load.pipe.filelist;
+
+import com.starrocks.load.pipe.PipeFileRecord;
+import com.starrocks.load.pipe.PipeId;
 
 import java.util.List;
