+++ fe/fe-core/src/main/java/com/starrocks/statistic/StatisticsMetaManager.java
@@ -30,7 +30,7 @@ import com.starrocks.common.Pair;
 import com.starrocks.common.UserException;
 import com.starrocks.common.util.FrontendDaemon;
 import com.starrocks.common.util.PropertyAnalyzer;
-import com.starrocks.load.pipe.FileListTableRepo;
+import com.starrocks.load.pipe.filelist.RepoCreator;
 import com.starrocks.qe.ConnectContext;
 import com.starrocks.server.GlobalStateMgr;
 import com.starrocks.server.RunMode;
@@ -357,7 +357,7 @@ public class StatisticsMetaManager extends FrontendDaemon {
         GlobalStateMgr.getCurrentAnalyzeMgr().clearStatisticFromDroppedTable();
         GlobalStateMgr.getCurrentAnalyzeMgr().clearExpiredAnalyzeStatus();
 
-        FileListTableRepo.RepoCreator.getInstance().run();
+        RepoCreator.getInstance().run();
     }
 
     public void createStatisticsTablesForTest() {
